requirements.txt
jinja2
stringcase
json-schema-codegen>=0.1.0
pyyaml